+++ service/controller/resource/organization/create.go
@@ -47,8 +47,8 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 
 	r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("created organization namespace %#q", orgNamespace.Name))
 
-	org.Status.Namespace = orgNamespace.Name
-	err = r.k8sClient.CtrlClient().Status().Update(ctx, &org)
+	patch := []byte(fmt.Sprintf(`{"status":{"namespace": "%s"}}`, orgNamespace.Name))
+	err = r.k8sClient.CtrlClient().Patch(ctx, &org, ctrl.RawPatch(types.MergePatchType, patch))
 	if err != nil {
 		return microerror.Mask(err)
 	}