+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.10.0] - 2021-10-12
+
 ### Added
 
 - Ensure `Organization` CR in Azure MCs have the `subscriptionid` annotation set.
@@ -87,7 +89,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First release.
 
-[Unreleased]: https://github.com/giantswarm/organization-operator/compare/v0.9.0...HEAD
+[Unreleased]: https://github.com/giantswarm/organization-operator/compare/v0.10.0...HEAD
+[0.10.0]: https://github.com/giantswarm/organization-operator/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/giantswarm/organization-operator/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/giantswarm/organization-operator/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/giantswarm/organization-operator/compare/v0.6.0...v0.7.0